another int/long fixup that was giving dodgy property values on amd64, shame on me for not reading xlibs (appauling) documentation more closely.
ok okan@

+++ app/cwm/screen.c
@@ -15,7 +15,7 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
- * $Id: screen.c,v 1.21 2009/12/10 23:14:58 oga Exp $
+ * $Id: screen.c,v 1.22 2009/12/11 17:57:38 oga Exp $
  */
 
 #include "headers.h"
@@ -102,7 +102,7 @@ screen_find_xinerama(struct screen_ctx *sc, int x, int y)
 void
 screen_update_geometry(struct screen_ctx *sc, int width, int height)
 {
-	int	geom[2];
+	long	geom[2];
 
 	sc->xmax = geom[0] = width;
 	sc->ymax = geom[1] = height;
